Introduce relay pipedream
This introduces a new set of GoCD pipelines generated from jsonnet.

The pipelines produced are:

- deploy-relay
- deploy-us
- deploy-monitor

`deploy-relay` is a simple pipeline that triggers a deployment of relay but doesn't
actually deploy anything.
`deploy-us` is the equivalent and deploy-relay and deploy-relay-pops that currently
exist. The only difference is that the pops pipeline has been turned into a stage.
`deploy-monitor` is a new pipeline that deploys to a single-tenant style environment.